refactor: split runtime setup modules for codacy
Extract runtime setup helper and UI modules so the setup entrypoint stays below Codacy's file-size threshold while preserving the existing public API and coverage contract.
